+++ calendra/america/__init__.py
@@ -1,0 +1,59 @@
+# -*- coding: utf-8 -*-
+from __future__ import (absolute_import, division, print_function,
+                        unicode_literals)
+
+from .brazil import (
+    Brazil, BrazilAcre, BrazilAlagoas, BrazilAmapa, BrazilAmazonas,
+    BrazilBahia, BrazilCeara, BrazilDistritoFederal, BrazilEspiritoSanto,
+    BrazilGoias, BrazilMaranhao, BrazilMatoGrosso, BrazilMatoGrossoDoSul,
+    BrazilPara, BrazilParaiba, BrazilPernambuco, BrazilPiaui,
+    BrazilRioDeJaneiro, BrazilRioGrandeDoNorte, BrazilRioGrandeDoSul,
+    BrazilRondonia, BrazilRoraima, BrazilSantaCatarina, BrazilSaoPauloState,
+    BrazilSaoPauloCity, BrazilSergipe, BrazilTocantins, BrazilVitoriaCity,
+    BrazilVilaVelhaCity, BrazilCariacicaCity, BrazilGuarapariCity,
+    BrazilSerraCity,
+)
+from .chile import Chile
+from .colombia import Colombia
+from .mexico import Mexico
+from .panama import Panama
+
+
+__all__ = (
+    Brazil,
+    BrazilAcre,
+    BrazilAlagoas,
+    BrazilAmapa,
+    BrazilAmazonas,
+    BrazilBahia,
+    BrazilCeara,
+    BrazilDistritoFederal,
+    BrazilEspiritoSanto,
+    BrazilGoias,
+    BrazilMaranhao,
+    BrazilMatoGrosso,
+    BrazilMatoGrossoDoSul,
+    BrazilPara,
+    BrazilParaiba,
+    BrazilPernambuco,
+    BrazilPiaui,
+    BrazilRioDeJaneiro,
+    BrazilRioGrandeDoNorte,
+    BrazilRioGrandeDoSul,
+    BrazilRondonia,
+    BrazilRoraima,
+    BrazilSantaCatarina,
+    BrazilSaoPauloState,
+    BrazilSaoPauloCity,
+    BrazilSergipe,
+    BrazilTocantins,
+    BrazilVitoriaCity,
+    BrazilVilaVelhaCity,
+    BrazilCariacicaCity,
+    BrazilGuarapariCity,
+    BrazilSerraCity,
+    Chile,
+    Colombia,
+    Mexico,
+    Panama,
+)