+++ calendra/europe/__init__.py
@@ -1,5 +1,9 @@
 # -*- coding: utf-8 -*-
+from .austria import Austria
 from .belgium import Belgium
+from .bulgaria import Bulgaria
+from .croatia import Croatia
+from .cyprus import Cyprus
 from .czech_republic import CzechRepublic
 from .estonia import Estonia
 from .european_central_bank import EuropeanCentralBank
@@ -9,17 +13,21 @@ from .france import France, FranceAlsaceMoselle
 from .greece import Greece
 from .hungary import Hungary
 from .iceland import Iceland
+from .ireland import Ireland
 from .italy import Italy
+from .latvia import Latvia
 from .luxembourg import Luxembourg
+from .malta import Malta
 from .netherlands import Netherlands
 from .norway import Norway
 from .poland import Poland
 from .portugal import Portugal
+from .romania import Romania
 from .slovakia import Slovakia
 from .slovenia import Slovenia
 from .spain import Spain, Catalonia
 from .sweden import Sweden
-from .switzerland import Switzerland
+from .switzerland import Switzerland, Vaud
 from .united_kingdom import UnitedKingdom, UnitedKingdomNorthernIreland
 
 # Germany
@@ -32,8 +40,12 @@ from .germany import (
 
 
 __all__ = (
+    Austria,
     Belgium,
+    Bulgaria,
     Catalonia,
+    Croatia,
+    Cyprus,
     CzechRepublic,
     Denmark,
     Estonia,
@@ -44,17 +56,22 @@ __all__ = (
     Greece,
     Hungary,
     Iceland,
+    Ireland,
     Italy,
+    Latvia,
     Luxembourg,
+    Malta,
     Netherlands,
     Norway,
     Poland,
     Portugal,
+    Romania,
     Slovakia,
     Slovenia,
     Spain,
     Sweden,
     Switzerland,
+    Vaud,
     UnitedKingdom,
     UnitedKingdomNorthernIreland,