+++ calendra/europe/slovakia.py
@@ -7,6 +7,10 @@ class Slovakia(WesternCalendar, ChristianMixin):
     include_epiphany = True
     include_easter_monday = True
     include_good_friday = True
+    include_all_saints = True
+    include_christmas_eve = True
+    include_boxing_day = True
+    boxing_day_label = "St. Stephen's Day (The Second Christmas Day)"
 
     FIXED_HOLIDAYS = WesternCalendar.FIXED_HOLIDAYS + (
         (1, 1, "Day of the Establishment of the Slovak Republic"),
@@ -16,9 +20,5 @@ class Slovakia(WesternCalendar, ChristianMixin):
         (8, 29, "Slovak National Uprising anniversary"),
         (9, 1, "Day of the Constitution of the Slovak Republic"),
         (9, 15, "Day of Blessed Virgin Mary, patron saint of Slovakia"),
-        (11, 1, "All Saints’ Day"),
         (11, 17, "Struggle for Freedom and Democracy Day"),
-        (12, 24, "Christmas Eve"),
-        (12, 25, "Christmas Day"),
-        (12, 26, "St. Stephen's Day (The Second Christmas Day)"),
     )