+++ calendra/tests/test_europe.py
@@ -1,36 +1,129 @@
 # coding=utf-8
 from datetime import date
+from collections import Counter
 
-from . import GenericCalendarTest
+from ..tests import GenericCalendarTest
+from ..europe import Austria
+from ..europe import Bulgaria
+from ..europe import Belgium
+from ..europe import Croatia
+from ..europe import Cyprus
 from ..europe import CzechRepublic
 from ..europe import Denmark
 from ..europe import Slovakia
-from ..europe import Estonia
 from ..europe import Finland
+from ..europe import Estonia
 from ..europe import Sweden
 from ..europe import France, FranceAlsaceMoselle
 from ..europe import Greece
 from ..europe import Hungary
 from ..europe import Iceland
+from ..europe import Ireland
 from ..europe import Italy
+from ..europe import Latvia
 from ..europe import Luxembourg
+from ..europe import Malta
 from ..europe import Netherlands
 from ..europe import Norway
 from ..europe import Poland
 from ..europe import Portugal
+from ..europe import Romania
 from ..europe import Spain, Catalonia
 from ..europe import Slovenia
+from ..europe import Switzerland, Vaud
 from ..europe import UnitedKingdom
 from ..europe import UnitedKingdomNorthernIreland
 from ..europe import EuropeanCentralBank
-from ..europe import Belgium
-from ..europe import Switzerland
-from ..europe import (
-    Germany, BadenWurttemberg, Bavaria, Berlin, Brandenburg, Bremen, Hamburg,
-    Hesse, MecklenburgVorpommern, LowerSaxony, NorthRhineWestphalia,
-    RhinelandPalatinate, Saarland, Saxony, SaxonyAnhalt, SchleswigHolstein,
-    Thuringia,
-)
+from ..europe import (Germany, BadenWurttemberg, Bavaria, Berlin,
+                                Brandenburg, Bremen, Hamburg, Hesse,
+                                MecklenburgVorpommern, LowerSaxony,
+                                NorthRhineWestphalia, RhinelandPalatinate,
+                                Saarland, Saxony, SaxonyAnhalt,
+                                SchleswigHolstein, Thuringia)
+
+
+class AustriaTest(GenericCalendarTest):
+    cal_class = Austria
+
+    def test_year_2016(self):
+        holidays = self.cal.holidays_set(2016)
+        self.assertIn(date(2016, 1, 1), holidays)  # New Years day
+        self.assertIn(date(2016, 1, 6), holidays)  # Epiphany
+        self.assertIn(date(2016, 3, 28), holidays)  # easter monday
+        self.assertIn(date(2016, 5, 1), holidays)  # National Holiday
+        self.assertIn(date(2016, 5, 5), holidays)  # Ascension Day
+        self.assertIn(date(2016, 5, 16), holidays)  # Whit monday
+        self.assertIn(date(2016, 5, 26), holidays)  # Corpus Christi
+        self.assertIn(date(2016, 8, 15), holidays)  # Assumption
+        self.assertIn(date(2016, 10, 26), holidays)  # national day again
+        self.assertIn(date(2016, 11, 1), holidays)  # all saints
+        self.assertIn(date(2016, 12, 8), holidays)  # Immaculate conception
+        self.assertIn(date(2016, 12, 25), holidays)  # Xmas
+        self.assertIn(date(2016, 12, 26), holidays)  # St Stephens
+
+
+class BulgariaTest(GenericCalendarTest):
+    cal_class = Bulgaria
+
+    def test_year_2016(self):
+        holidays = self.cal.holidays_set(2016)
+        self.assertIn(date(2016, 1, 1), holidays)   # New Year's Day
+        self.assertIn(date(2016, 3, 3), holidays)   # Liberation Day
+        self.assertIn(date(2016, 3, 27), holidays)   # Easter Sun
+        self.assertIn(date(2016, 3, 28), holidays)   # Easter Mon
+        self.assertIn(date(2016, 5, 1), holidays)   # International Workers'
+        self.assertIn(date(2016, 5, 6), holidays)   # St George's Day
+        self.assertIn(date(2016, 5, 24), holidays)   # St Cyril & Methodius
+        self.assertIn(date(2016, 9, 6), holidays)   # Unification Day
+        self.assertIn(date(2016, 9, 22), holidays)   # Independence Day
+        self.assertIn(date(2016, 12, 24), holidays)   # Christmas Eve
+        self.assertIn(date(2016, 12, 25), holidays)   # Christmas 1
+        self.assertIn(date(2016, 12, 26), holidays)   # Christmas 2
+        # Non-attendance day for schools, otherwise a working day.
+        self.assertNotIn(date(2016, 11, 1), holidays)   # National Awakening
+
+
+class CroatiaTest(GenericCalendarTest):
+    cal_class = Croatia
+
+    def test_year_2016(self):
+        holidays = self.cal.holidays_set(2016)
+        self.assertIn(date(2016, 1, 1), holidays)   # New Year's Day Nova Godin
+        self.assertIn(date(2016, 1, 6), holidays)   # Epiphany Bogojavljenje,
+        self.assertIn(date(2016, 3, 27), holidays)  # Easter Sunday Uskrs i us
+        self.assertIn(date(2016, 3, 28), holidays)  # Easter Monday
+        self.assertIn(date(2016, 5, 1), holidays)   # Intl Workers' Day Međunar
+        self.assertIn(date(2016, 5, 26), holidays)  # Corpus Christi Tijelovo
+        self.assertIn(date(2016, 6, 22), holidays)  # Anti-Fascist Day Dan anti
+        self.assertIn(date(2016, 6, 25), holidays)  # Statehood Day 	Dan drž
+        self.assertIn(date(2016, 8, 5), holidays)   # Victory & Homeland Thanks
+        self.assertIn(date(2016, 8, 15), holidays)  # Assumption of Mary 	Vel
+        self.assertIn(date(2016, 10, 8), holidays)  # Independence Day Dan neov
+        self.assertIn(date(2016, 11, 1), holidays)  # All Saints' Day Dan svih
+        self.assertIn(date(2016, 12, 25), holidays)  # Christmas Božić
+        self.assertIn(date(2016, 12, 26), holidays)  # St. Stephen's Day Prvi d
+
+
+class Cyprus(GenericCalendarTest):
+    cal_class = Cyprus
+
+    def test_year_2017(self):
+        holidays = self.cal.holidays_set(2017)
+        self.assertIn(date(2017, 1, 1), holidays)    # New Year's Day
+        self.assertIn(date(2017, 1, 6), holidays)    # Epiphany
+        self.assertIn(date(2017, 2, 27), holidays)   # Green Monday
+        self.assertIn(date(2017, 3, 25), holidays)   # Greek Independence Day
+        self.assertIn(date(2017, 4, 1), holidays)    # Cyprus National Holiday
+        self.assertIn(date(2017, 4, 14), holidays)   # Good Friday (Orthodox)
+        self.assertIn(date(2017, 4, 17), holidays)   # Easter Monday (Orthodox)
+        self.assertIn(date(2017, 4, 18), holidays)   # Easter Tues banks only
+        self.assertIn(date(2017, 5, 1), holidays)    # Labour Day/May Day
+        self.assertIn(date(2017, 6, 5), holidays)    # Orthodox Pentecost Mon
+        self.assertIn(date(2017, 7, 15), holidays)   # Assumption
+        self.assertIn(date(2017, 10, 1), holidays)   # Cyprus Independence Day
+        self.assertIn(date(2017, 10, 28), holidays)  # Ochi Day
+        self.assertIn(date(2017, 12, 25), holidays)  # Christmas Day
+        self.assertIn(date(2017, 12, 26), holidays)  # Boxing Day
 
 
 class CzechRepublicTest(GenericCalendarTest):
@@ -102,6 +195,19 @@ class SlovakiaTest(GenericCalendarTest):
         self.assertIn(date(2013, 12, 24), holidays)
         self.assertIn(date(2013, 12, 25), holidays)
         self.assertIn(date(2013, 12, 26), holidays)
+
+    def test_removed_duplicate(self):
+        holidays = self.cal.holidays(2017)
+        counter = Counter(day for day, label in holidays)
+        # Only one "all saints"
+        self.assertEqual(counter[date(2017, 11, 1)], 1)
+        # Only one "XMas eve"
+        self.assertEqual(counter[date(2017, 12, 24)], 1)
+        # De-duplicate, XMas day was configured twice.
+        # Refs #205.
+        self.assertEqual(counter[date(2017, 12, 25)], 1)
+        # Only one St Stephen
+        self.assertEqual(counter[date(2017, 12, 25)], 1)
 
 
 class SwedenTest(GenericCalendarTest):
@@ -303,6 +409,41 @@ class HungaryTest(GenericCalendarTest):
         self.assertIn(date(2013, 12, 25), holidays)  # XMas
         self.assertIn(date(2013, 12, 26), holidays)  # Second day of XMas
 
+    def test_year_good_friday(self):
+        # Since the year 2017, Good Friday is a non-working day un Hungary
+        # Refs #203
+        holidays = self.cal.holidays_set(2016)
+        good_friday_2016 = date(2016, 3, 25)
+        self.assertNotIn(good_friday_2016, holidays)
+        # 2017
+        holidays = self.cal.holidays_set(2017)
+        good_friday_2017 = date(2017, 4, 14)
+        self.assertIn(good_friday_2017, holidays)
+
+
+class MaltaTest(GenericCalendarTest):
+    """Rollover rules changed in 2005"""
+    cal_class = Malta
+
+    def test_year_2017(self):
+        holidays = self.cal.holidays_set(2017)
+        # National Holidays
+        self.assertIn(date(2017, 3, 31), holidays)  # Jum il-Ħelsien
+        self.assertIn(date(2017, 6, 7), holidays)  # Sette Giugno
+        self.assertIn(date(2017, 9, 8), holidays)  # Jum il-Vitorja
+        self.assertIn(date(2017, 9, 21), holidays)  # Jum l-Indipendenza
+        self.assertIn(date(2017, 12, 13), holidays)  # Jum ir-Repubblika
+        # Public Holidays
+        self.assertIn(date(2017, 1, 1), holidays)  # L-Ewwel tas-Sena
+        self.assertIn(date(2017, 2, 10), holidays)  # Nawfraġju ta' San Pawl
+        self.assertIn(date(2017, 3, 19), holidays)  # San Ġużepp
+        self.assertIn(date(2017, 4, 14), holidays)  # Il-Ġimgħa l-Kbira
+        self.assertIn(date(2017, 5, 1), holidays)  # Jum il-Ħaddiem
+        self.assertIn(date(2017, 6, 29), holidays)  # L-Imnarja
+        self.assertIn(date(2017, 8, 15), holidays)  # Santa Marija
+        self.assertIn(date(2017, 12, 8), holidays)  # Il-Kunċizzjoni
+        self.assertIn(date(2017, 12, 25), holidays)  # Il-Milied
+
 
 class NorwayTest(GenericCalendarTest):
     cal_class = Norway
@@ -364,6 +505,66 @@ class IcelandTest(GenericCalendarTest):
         self.assertIn(date(2013, 12, 31), holidays)
 
 
+class IrelandTest(GenericCalendarTest):
+    cal_class = Ireland
+
+    def test_year_2013(self):
+        holidays = self.cal.holidays_set(2013)
+        self.assertIn(date(2013, 1, 1), holidays)    # Tue New Year's Day
+        self.assertIn(date(2013, 3, 17), holidays)   # Sun St Patricks day
+        self.assertIn(date(2013, 3, 18), holidays)   # Mon St Patricks shift
+        self.assertIn(date(2013, 4, 1), holidays)    # Mon Easter Monday
+        self.assertIn(date(2013, 5, 6), holidays)    # Mon May day
+        self.assertIn(date(2013, 6, 3), holidays)    # Mon June Holiday
+        self.assertIn(date(2013, 8, 5), holidays)    # Mon August Holiday
+        self.assertIn(date(2013, 10, 28), holidays)  # Mon October Holiday
+        self.assertIn(date(2013, 12, 25), holidays)  # Wed Christmas
+        self.assertIn(date(2013, 12, 26), holidays)  # Thu St Stepehen's
+
+    def test_shift_2012(self):
+        holidays = self.cal.holidays_set(2012)
+        self.assertIn(date(2012, 1, 1), holidays)    # new year day
+        self.assertIn(date(2012, 1, 2), holidays)    # new year day shift
+
+    def test_shift_2011(self):
+        holidays = self.cal.holidays_set(2011)
+        self.assertIn(date(2011, 12, 25), holidays)  # Christmas it's sunday
+        self.assertIn(date(2011, 12, 26), holidays)  # Xmas day shift
+        self.assertIn(date(2011, 12, 27), holidays)  # St Stephen's day shift
+
+    def test_shift_2015(self):
+        holidays = self.cal.holidays_set(2015)
+        self.assertIn(date(2015, 12, 25), holidays)  # Christmas it's friday
+        self.assertIn(date(2015, 12, 26), holidays)  # St. Stephen's day Sat
+        self.assertIn(date(2015, 12, 28), holidays)  # St. Stephen's day shift
+
+    def test_whit_mon(self):
+        holidays_1972 = self.cal.holidays_set(1972)
+        holidays_1973 = self.cal.holidays_set(1973)
+        # holidays_1974 = self.cal.holidays_set(1974) Overlaps with June Hol
+        holidays_1975 = self.cal.holidays_set(1975)
+        self.assertIn(date(1972, 5, 22), holidays_1972)     # Whit Monday
+        self.assertIn(date(1973, 6, 11), holidays_1973)
+        # self.assertNotIn(date(1974, 6, 3), holidays_1974) # Overlaps Jun Hol
+        self.assertNotIn(date(1975, 5, 19), holidays_1975)  # No Whit Monday
+
+    def test_may_day(self):
+        holidays_1993 = self.cal.holidays_set(1993)
+        holidays_1994 = self.cal.holidays_set(1994)
+        holidays_1995 = self.cal.holidays_set(1995)
+        self.assertNotIn(date(1993, 5, 1), holidays_1993)  # No May Day
+        self.assertIn(date(1994, 5, 2), holidays_1994)     # May Day
+        self.assertIn(date(1995, 5, 1), holidays_1995)
+
+    def test_october_holiday(self):
+        holidays_1976 = self.cal.holidays_set(1976)
+        holidays_1977 = self.cal.holidays_set(1977)
+        holidays_1978 = self.cal.holidays_set(1978)
+        self.assertNotIn(date(1976, 10, 25), holidays_1976)  # No October Hol
+        self.assertIn(date(1977, 10, 31), holidays_1977)     # October Hol
+        self.assertIn(date(1978, 10, 30), holidays_1978)
+
+
 class ItalyTest(GenericCalendarTest):
     cal_class = Italy
 
@@ -380,6 +581,35 @@ class ItalyTest(GenericCalendarTest):
         self.assertIn(date(2013, 12, 8), holidays)  # immaculate Conception
         self.assertIn(date(2013, 12, 25), holidays)  # christmas
         self.assertIn(date(2013, 12, 26), holidays)  # San Stefano
+
+
+class LatviaTest(GenericCalendarTest):
+
+    cal_class = Latvia
+
+    def test_year_2017(self):
+        """
+        https://www.bank.lv/en/about-us/public-holidays-in-latvia
+        """
+        holidays = self.cal.holidays_set(2017)
+        self.assertIn(date(2017, 1, 1), holidays)  # New Year's Day
+        self.assertIn(date(2017, 4, 14), holidays)  # Easter 14 and 17 april
+        self.assertIn(date(2017, 4, 16), holidays)  # Easter 14 and 17 april
+        self.assertIn(date(2017, 4, 17), holidays)  # Easter 14 and 17 april
+        self.assertIn(date(2017, 5, 1), holidays)  # Labour Day, C
+        self.assertIn(date(2017, 5, 4), holidays)  # Restoration of Independe
+        self.assertIn(date(2017, 6, 23), holidays)  # Midsummer Day
+        self.assertIn(date(2017, 6, 24), holidays)  # St. John's Day
+        self.assertIn(date(2017, 11, 18), holidays)  # Proclamation
+        self.assertIn(date(2017, 11, 20), holidays)  # Proclamation Observed
+        self.assertIn(date(2017, 12, 24), holidays)  # Christmas
+        self.assertIn(date(2017, 12, 25), holidays)  # Christmas
+        self.assertIn(date(2017, 12, 26), holidays)  # Christmas
+        self.assertIn(date(2017, 12, 31), holidays)  # New Year's Eve
+
+    def test_year_2013(self):
+        holidays = self.cal.holidays_set(2013)
+        self.assertIn(date(2013, 5, 6), holidays)  # Restoration Day Observed
 
 
 class LuxembourgTest(GenericCalendarTest):
@@ -427,6 +657,34 @@ class NetherlandsTest(GenericCalendarTest):
         """ In 1990 Queen's day was on 30 April """
         holidays = self.cal.holidays_set(1990)
         self.assertIn(date(1990, 4, 30), holidays)   # Queen's Day
+
+    def test_new_years_eve(self):
+        # For some reason, the new year's eve was added to the list of fixed
+        # holidays. It appears it's not a holiday
+        holidays = self.cal.holidays_set(2016)
+        self.assertNotIn(date(2016, 12, 31), holidays)
+
+
+class Romania(GenericCalendarTest):
+    cal_class = Romania
+
+    def test_year_2017(self):
+        holidays = self.cal.holidays_set(2017)
+        self.assertIn(date(2017, 1, 1), holidays)  # Anul Nou New Year's Day
+        self.assertIn(date(2017, 1, 2), holidays)  # Anul Nou Day after New Yr
+        self.assertIn(date(2017, 1, 24), holidays)  # Unirea Principatelor Rom
+        self.assertIn(date(2017, 4, 14), holidays)  # Orthodox Good Fri
+        self.assertIn(date(2017, 4, 16), holidays)  # Orthodox Easter Sun
+        self.assertIn(date(2017, 4, 17), holidays)  # Orthodox Easter Mon
+        self.assertIn(date(2017, 5, 1), holidays)  # Ziua Muncii Labour Day
+        self.assertIn(date(2017, 6, 1), holidays)  # Ziua Copilului Children's
+        self.assertIn(date(2017, 6, 4), holidays)  # Pentecost
+        self.assertIn(date(2017, 6, 5), holidays)  # Whit Monday
+        self.assertIn(date(2017, 8, 15), holidays)  # Adormirea Maicii Domnului
+        self.assertIn(date(2017, 11, 30), holidays)  # Sfântul Andrei St. Andre
+        self.assertIn(date(2017, 12, 1), holidays)  # Ziua Națională/Marea Unir
+        self.assertIn(date(2017, 12, 25), holidays)  # Crăciunul Christmas
+        self.assertIn(date(2017, 12, 26), holidays)  # Crăciunul Christmas
 
 
 class UnitedKingdomTest(GenericCalendarTest):
@@ -551,6 +809,18 @@ class GermanyTest(GenericCalendarTest):
         self.assertIn(date(2015, 10, 3), holidays)
         self.assertIn(date(2015, 12, 25), holidays)
         self.assertIn(date(2015, 12, 26), holidays)
+
+    def test_year_reformation_day(self):
+        # Reformation Day is included each year in:
+        # Brandenburg, MecklenburgVorpommern, Saxony, SaxonyAnhalt & Thuringia
+        holidays = self.cal.holidays_set(2015)
+        if self.cal.include_reformation_day:
+            self.assertIn(date(2015, 10, 31), holidays)
+        else:
+            self.assertNotIn(date(2015, 10, 31), holidays)
+        # But in the year 2017, it's included for the whole country
+        holidays = self.cal.holidays_set(2017)
+        self.assertIn(date(2017, 10, 31), holidays)
 
 
 class BadenWurttembergTest(GermanyTest):
@@ -734,6 +1004,12 @@ class SaxonyTest(GermanyTest):
         self.assertIn(date(2015, 11, 18), holidays)
         self.assertIn(date(2015, 10, 31), holidays)
 
+    def test_extra_2016(self):
+        holidays = self.cal.holidays_set(2016)
+        self.assertIn(date(2016, 11, 16), holidays)
+        self.assertNotIn(date(2016, 11, 23), holidays)
+        self.assertIn(date(2016, 10, 31), holidays)
+
 
 class SaxonyAnhaltTest(GermanyTest):
     cal_class = SaxonyAnhalt
@@ -895,6 +1171,32 @@ class SloveniaTest(GenericCalendarTest):
         self.assertIn(date(2016, 12, 25), holidays)
         self.assertIn(date(2016, 12, 26), holidays)
 
+    def test_january_second(self):
+        # From 1955 until May 2012, when the National Assembly of Slovenia
+        # passed the Public Finance Balance Act, 2 January was a work-free day.
+        # It has been re-introduced in 2017.
+        # Source - Wikipedia
+        # https://en.wikipedia.org/wiki/Public_holidays_in_Slovenia
+
+        # Before 1955...
+        holidays = self.cal.holidays_set(1953)
+        self.assertNotIn(date(1953, 1, 2), holidays)
+        holidays = self.cal.holidays_set(1954)
+        self.assertNotIn(date(1953, 1, 2), holidays)
+        # Ranging from 1955 to 2012
+        for year in range(1955, 2013):
+            holidays = self.cal.holidays_set(year)
+            self.assertIn(date(year, 1, 2), holidays)
+
+        # As of 2012, it was no more
+        for year in range(2013, 2017):
+            holidays = self.cal.holidays_set(year)
+            self.assertNotIn(date(year, 1, 2), holidays)
+
+        # In 2017, january 2nd became a state holiday
+        holidays = self.cal.holidays_set(2017)
+        self.assertIn(date(2017, 1, 2), holidays)
+
 
 class SwitzerlandTest(GenericCalendarTest):
     cal_class = Switzerland
@@ -927,6 +1229,22 @@ class SwitzerlandTest(GenericCalendarTest):
         self.assertIn(date(2016, 8, 1), holidays)
         self.assertIn(date(2016, 12, 25), holidays)
         self.assertIn(date(2016, 12, 26), holidays)
+
+
+class VaudTest(GenericCalendarTest):
+    cal_class = Vaud
+
+    def test_year_2016(self):
+        holidays = self.cal.holidays_set(2016)
+        self.assertIn(date(2016, 9, 19), holidays)
+        self.assertNotIn(date(2016, 5, 1), holidays)
+        self.assertNotIn(date(2016, 12, 26), holidays)
+
+    def test_year_2017(self):
+        holidays = self.cal.holidays_set(2017)
+        self.assertIn(date(2017, 9, 18), holidays)
+        self.assertNotIn(date(2017, 5, 1), holidays)
+        self.assertNotIn(date(2017, 12, 26), holidays)
 
 
 class EstoniaTest(GenericCalendarTest):